refactor(library/typeUtilities): moves "/Boolean.ts" to "boolean/Not.ts"
