+++ src/library/typeUtilities/boolean/Not.ts
